Migrate PointerLock to dom
